+++ material/src/components/verticalDragNDropSelection/index.ts
@@ -1,0 +1,1 @@
+export * from './verticalDragNDropSelection.interface';