+++ material/src/components/verticalDragNDropSelection/verticalDragNDropSelection.interface.ts
@@ -1,0 +1,17 @@
+/**
+ * Css classes for vertical drag n drop selection
+ */
+export interface CssClassesVerticalDragNDropSelection
+{
+    titleClass?: string;
+    containerClass?: string;
+    itemClass?: string;
+}
+
+/**
+ * Texts that are used within VerticalDragNDropSelection
+ */
+export interface VerticalDragNDropSelectionTexts
+{
+    selectionTitle?: string;
+}